New and old API changes
Mostly worked on creating a new API route for creating new ingredient entries and new user ingredient entries. This requires further validation such as verifying if the user already does have said ingredient before adding it to the database, and if does it should update the entry, but will work on it later.